pyproject.toml
[project]
name = "test-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.84.0",
    "flask>=3.1.3",
    "python-dotenv>=1.2.1",
]
